UPDATE:TASK-AUTHORIZATIONS

+++ src/tasks/tasks.service.ts
@@ -16,22 +16,24 @@ export class TasksService {
 		private taskRepository: TaskRepository,
 	) {}
 
-	async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-		return await this.taskRepository.getTasks(filterDto);
+	async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+		return await this.taskRepository.getTasks(filterDto, user);
 	}
 
-	async getTaskById(id: number): Promise<Task> {
-		const task = await this.taskRepository.findOneBy({ id });
+	async getTaskById(id: number, user: User): Promise<Task> {
+		const task = await this.taskRepository.findOne({
+			where: { id, userId: user.id },
+		});
 		if (!task) {
 			throw new NotFoundException();
 		}
 		return task;
 	}
 
-	async deleteTaskById(id: number): Promise<void> {
-		const result = await this.taskRepository.delete(id);
+	async deleteTaskById(id: number, user: User): Promise<void> {
+		const result = await this.taskRepository.delete({ id, userId: user.id });
 		if (!result.affected) {
-			throw new NotFoundException();
+			throw new NotFoundException(`Task with ID '${id}' not found`);
 		}
 	}
 
@@ -39,8 +41,12 @@ export class TasksService {
 		return this.taskRepository.createTask(createTaskDto, user);
 	}
 
-	async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-		const task = await this.getTaskById(id);
+	async updateTaskStatus(
+		id: number,
+		status: TaskStatus,
+		user: User,
+	): Promise<Task> {
+		const task = await this.getTaskById(id, user);
 		task.status = status;
 		await task.save();
 		return task;
